requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
pydantic==2.9.2
feedparser==6.0.11
trafilatura==1.9.0
beautifulsoup4==4.12.3
html5lib==1.1
httpx==0.27.2
python-dateutil==2.9.0.post0
rapidfuzz==3.9.6
PyYAML==6.0.2
openai==1.47.0
